du_runmeng/Non-learning-main/Old/requirements.txt
boto3==1.16.30
botocore==1.19.30
certifi==2020.11.8
chardet==3.0.4
Cython==3.0.0
dataclasses==0.7
docopt==0.6.2
future==0.18.2
galois==0.0.26
gmpy2==2.1.5
idna==2.10
jmespath==0.10.0
llvmlite==0.36.0
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
mpmath==1.3.0
numba==0.53.1
numpy==1.19.5
olefile==0.44
pandas==1.1.4
phe==1.5.0
Pillow @
pipreqs==0.4.11
PyFlint==0.1
python-dateutil==2.8.1
pytorch-pretrained-bert==0.6.2
pytz==2020.4
regex==2020.11.13
requests==2.25.0
s3transfer==0.3.3
scipy==1.5.4
six==1.10.0
sympy==1.9
torch==1.7.0
torchaudio==0.7.0
torchvision==0.8.1
tqdm==4.54.1
typing-extensions
urllib3==1.26.2
wincertstore==0.2
yarg==0.1.9
